--- a/api/src/main/java/graphql/nadel/NadelExecutionHints.java
+++ b/api/src/main/java/graphql/nadel/NadelExecutionHints.java
@@ -11,10 +11,12 @@ public class NadelExecutionHints {
     private NadelExecutionHints(Builder builder) {
         this.transformsOnHydrationFields = builder.transformsOnHydrationFields;
         this.legacyOperationNames = builder.legacyOperationNames;
+        this.newJsonNodeTraversal = builder.newJsonNodeTraversal;
     }
 
     private final LegacyOperationNamesHint legacyOperationNames;
     private final boolean transformsOnHydrationFields;
+    private final boolean newJsonNodeTraversal;
 
     /**
      * Flag to determine whether nextgen will generate the traditional nadel_2_service_opName
@@ -33,6 +35,13 @@ public class NadelExecutionHints {
      */
     public boolean getTransformsOnHydrationFields() {
         return transformsOnHydrationFields;
+    }
+
+    /**
+     * Flag to use JsonNodes with traversal caching over the slower but battle-tested JsonNodeExtractor.
+     */
+    public boolean getNewJsonNodeTraversal() {
+        return newJsonNodeTraversal;
     }
 
     /**
@@ -57,6 +66,7 @@ public class NadelExecutionHints {
 
     public static class Builder {
         private boolean transformsOnHydrationFields;
+        private boolean newJsonNodeTraversal;
         private LegacyOperationNamesHint legacyOperationNames = service -> false;
 
         private Builder() {
@@ -68,6 +78,14 @@ public class NadelExecutionHints {
 
         public Builder transformsOnHydrationFields(boolean flag) {
             this.transformsOnHydrationFields = flag;
+            return this;
+        }
+
+        /**
+         * @see NadelExecutionHints#getNewJsonNodeTraversal()
+         */
+        public Builder newJsonNodeTraversal(boolean flag) {
+            this.newJsonNodeTraversal = flag;
             return this;
         }