Did a dig at performance last night and got a bunch of performance improvements which brings down the execution time by a lot.

Old engine

```
Result "benchmark.LargeRenamedResponseBenchmark.benchMarkAvgTime":
  53.688 ±(99.9%) 2.307 ms/op [Average]
  (min, avg, max) = (50.883, 53.688, 57.167), stdev = 2.158
  CI (99.9%): [51.381, 55.995] (assumes normal distribution)
```

New engine without changes

```
Result "benchmark.LargeRenamedResponseBenchmark.benchMarkAvgTime":
  42.152 ±(99.9%) 1.423 ms/op [Average]
  (min, avg, max) = (40.100, 42.152, 44.873), stdev = 1.331
  CI (99.9%): [40.730, 43.575] (assumes normal distribution)
```

With caching traversals using `JsonNodes` instead of `JsonNodeExtractor`

```
Result "benchmark.LargeRenamedResponseBenchmark.benchMarkAvgTime":
  35.301 ±(99.9%) 0.846 ms/op [Average]
  (min, avg, max) = (33.777, 35.301, 36.819), stdev = 0.792
  CI (99.9%): [34.455, 36.147] (assumes normal distribution)
```

After removing `DataMutations`

```
Result "benchmark.LargeRenamedResponseBenchmark.benchMarkAvgTime":
  27.076 ±(99.9%) 0.998 ms/op [Average]
  (min, avg, max) = (25.917, 27.076, 28.899), stdev = 0.934
  CI (99.9%): [26.078, 28.074] (assumes normal distribution)
```

After updating `JsonMap.getAt` implementation

```
Result "benchmark.LargeRenamedResponseBenchmark.benchMarkAvgTime":
  20.348 ±(99.9%) 0.914 ms/op [Average]
  (min, avg, max) = (18.878, 20.348, 22.288), stdev = 1.053
  CI (99.9%): [19.434, 21.262] (assumes normal distribution)
```

I still need to change `NadelCoerceTransform` to stop running all the time. It effectively runs for every leaf node, which is a decent performance hit. I tested straight up removing the transform and got a drop of ~3 ms in performance.

There's probably still some areas where we can improve this. This is by no means done either.